+++ server.h
@@ -1,8 +1,8 @@
 #ifndef SERVER_H
 #define SERVER_H
 
-int accept_connection(int listensd, struct sockaddr_in * cliaddr, unsigned int * cliseq);
-void complete_handshake(unsigned int * cliseq, struct qnode ** send_queue);
+int accept_connection(int listensd, struct sockaddr_in * cliaddr, unsigned long * cliseq);
+void complete_handshake(unsigned long * cliseq, struct qnode ** send_queue);
 void * send_message(void * args);
 void send_list(struct qnode ** send_queue);
 void * msg_handler(void * args);
